Formatting.

+++ mlx_vlm/models/bonsai/scheduler.py
@@ -23,7 +23,9 @@ class FlowMatchEulerDiscreteScheduler:
         return float(a * num_steps + b)
 
     @staticmethod
-    def _time_shift_exponential_array(mu: float, sigma_power: float, t: mx.array) -> mx.array:
+    def _time_shift_exponential_array(
+        mu: float, sigma_power: float, t: mx.array
+    ) -> mx.array:
         return mx.exp(mu) / (mx.exp(mu) + ((1.0 / t - 1.0) ** sigma_power))
 
     @classmethod
@@ -34,13 +36,19 @@ class FlowMatchEulerDiscreteScheduler:
         num_inference_steps: int,
         num_train_timesteps: int = 1000,
     ) -> tuple[mx.array, mx.array]:
-        sigmas = mx.linspace(1.0, 1.0 / num_inference_steps, num_inference_steps, dtype=mx.float32)
-        mu = cls._compute_empirical_mu(image_seq_len=image_seq_len, num_steps=num_inference_steps)
+        sigmas = mx.linspace(
+            1.0, 1.0 / num_inference_steps, num_inference_steps, dtype=mx.float32
+        )
+        mu = cls._compute_empirical_mu(
+            image_seq_len=image_seq_len, num_steps=num_inference_steps
+        )
         sigmas = cls._time_shift_exponential_array(mu, 1.0, sigmas)
         timesteps = sigmas * num_train_timesteps
         sigmas = mx.concatenate([sigmas, mx.zeros((1,), dtype=sigmas.dtype)], axis=0)
         return timesteps, sigmas
 
     def step(self, *, noise: mx.array, step_index: int, latents: mx.array) -> mx.array:
-        dt = (self.sigmas[step_index + 1] - self.sigmas[step_index]).astype(latents.dtype)
+        dt = (self.sigmas[step_index + 1] - self.sigmas[step_index]).astype(
+            latents.dtype
+        )
         return latents + dt * noise.astype(latents.dtype)
